Removing unused code. Diff looks a bit weird since files were renamed (removed prefixed).
